Revert "Исправление типа данных в методе findAllFilmByDirector на Integer"
This reverts commit b88ef6fe8cf881ebcca75fc784ec17ec9a8cf7de.

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -71,7 +71,7 @@ public class FilmController {
     }
 
     @GetMapping("/director/{directorId}")
-    public List<Film> findAllFilmByDirector(@PathVariable Integer directorId,
+    public List<Film> findAllFilmByDirector(@PathVariable Long directorId,
                                             @RequestParam("sortBy") String sortBy) {
         return filmService.findAllFilmByDirector(directorId, sortBy);
     }
